Trying to get tutorial to populate from JSON stored in a service.
Two problems:

Major:  Can't get images to render.
Minor:  Can't get toggle button contstructed.

Both problems seems to involve WHEN the code gets rendered.   I think that initial rendering gets the code out of the JSON, but that's it.

THought is to pass it to a custom directive, render it, then pass it back.